fix: Create forfeit TXs correctly when leaves are in multiple levels
When we were going against Arkd v0.6.X, the `concurrent_boarding` test
would always manage to create all the forfeit transactions needed to
refresh the VTXOs sent between Alice, Bob and Claire.

With the upgrade to patch `6246f89716f7795801c2d453a02ec911898c4e15`
of arkd, this test would now fail because we were unable to find all
the necessary connector outputs to construct all the forfeit
transactions. Our implementation assumed that all connector outputs
would be in the last level of the connector tree, but of course leaves
can appear in higher levels too.

I suspect the arkd upgrade changed the way the connector outputs tree
is built, which is surprising.